Reorganised so that the module structure now mirrors the pandas ones it is overriding
