+++ anesthetic/plotting/_matplotlib/__init__.py
@@ -7,23 +7,23 @@ from pandas.plotting._matplotlib import (  # noqa: F401
         register,
         deregister
 )
-from anesthetic._matplotlib.boxplot import (  # noqa: F401
+from anesthetic.plotting._matplotlib.boxplot import (  # noqa: F401
     BoxPlot,
     boxplot,
     boxplot_frame,
 )
 from pandas.plotting._matplotlib import boxplot_frame_groupby  # noqa: F401
-from anesthetic._matplotlib.core import (  # noqa: F401
+from anesthetic.plotting._matplotlib.core import (  # noqa: F401
     ScatterPlot,
     HexBinPlot,
 )
-from anesthetic._matplotlib.hist import (  # noqa: F401
+from anesthetic.plotting._matplotlib.hist import (  # noqa: F401
     HistPlot,
     KdePlot,
     hist_frame,
     hist_series,
 )
-from anesthetic._matplotlib.misc import (  # noqa: F401
+from anesthetic.plotting._matplotlib.misc import (  # noqa: F401
     bootstrap_plot,
     scatter_matrix,
 )
